Refactor: 공통 클래스를 오브젝트로 변경
